Extracted low-level file handling from TenantResourceManager out with an interface
